requirements.txt
# ============================================================
# JARVIS AI Assistant - Python Dependencies
# ============================================================

# ============================================================
# CORE DEPENDENCIES (Required)
# ============================================================
cryptography>=41.0.0
bcrypt>=4.0.0

# ============================================================
# AI INTELLIGENCE (Choose one approach)
# ============================================================

# Option 1: Ollama (RECOMMENDED - Best quality, streaming, internet access)
ollama>=0.1.0
duckduckgo-search>=4.0.0
pyyaml>=6.0

# Option 2: Transformers + GPT4All (Fallback - works offline)
# Uncomment if not using Ollama:
# transformers>=4.35.0
# torch>=2.0.0
# gpt4all>=2.0.0

# ============================================================
# COMPUTER VISION (Optional - for camera/object detection)
# ============================================================
# Uncomment to install vision features:
opencv-python>=4.8.0
ultralytics>=8.0.0

# ============================================================
# 3D RENDERING (Optional - for material visualization)
# ============================================================
# Uncomment to install 3D features:
pyvista>=0.43.0
matplotlib>=3.7.0

# ============================================================
# TEXT-TO-SPEECH (Recommended)
# ============================================================
pyttsx3>=2.90

# ============================================================
# OPTIONAL ENHANCEMENTS
# ============================================================
# Faster Whisper for voice recognition:
faster-whisper>=0.9.0

# Face recognition:
face-recognition>=1.3.0

# Audio processing:
pyaudio>=0.2.13

# ============================================================
# NOTES
# ============================================================
# 1. Install core only:
#    pip install -r requirements.txt
#
# 2. Install everything (uncomment all above):
#    pip install -r requirements.txt
#
# 3. Install specific features:
#    pip install cryptography bcrypt pyttsx3 transformers torch
#
# 4. Some packages require Visual Studio C++ Build Tools
#    Download: https://visualstudio.microsoft.com/downloads/
#
# 5. GPU acceleration (optional):
#    pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118